+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Copy selected date to clipboard.
+- "y"       = copy selected date to clipboard
+- "b" / "H" = back to start of week
+- "e" / "L" = end of week
+- "w"       = start of next week
 
 ## [0.1.0]
 ### Added